MALI: rockchip: upgrade bifrost DDK to g22p0-01eac0, from g21p0-01eac0

+++ drivers/gpu/arm/bifrost/tests/kutf/kutf_utils.c
@@ -37,7 +37,7 @@ const char *kutf_dsprintf(struct kutf_mempool *pool, const char *fmt, ...)
 {
 	va_list args;
 	int len;
-	int size;
+	size_t size;
 	void *buffer;
 
 	mutex_lock(&buffer_lock);
@@ -54,7 +54,7 @@ const char *kutf_dsprintf(struct kutf_mempool *pool, const char *fmt, ...)
 		pr_warn("%s: Truncated dsprintf message %s\n", __func__, fmt);
 		size = sizeof(tmp_buffer);
 	} else {
-		size = len + 1;
+		size = (size_t)(len + 1);
 	}
 
 	buffer = kutf_mempool_alloc(pool, size);
